Change the return type of Configurator::InitialDelay from int to double.
This allows the Configurator::InitialDelay function to return sub-second
values for unit-testing and functional-testing of //chrome/updater.

The code in the CUS service was changed to convert from double
to a TimeDelta. Also, the Configurator code was changed to parse
a literal to double.

Where possible, the type was changed from int
to double. In general, the standard conversion from int to double
is used (for instance in unit tests). It's possible that there might
be cases where a truncation from double to int occurs (because these
call sites were missed) but this should not create an issue at
runtime because the values involved in the initial delay are small.

Bug: 1144151
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2635022
Commit-Queue: Sorin Jianu <sorin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#845014}

+++ chrome/browser/component_updater/chrome_component_updater_configurator.cc
@@ -54,7 +54,7 @@ class ChromeConfigurator : public update_client::Configurator {
                      PrefService* pref_service);
 
   // update_client::Configurator overrides.
-  int InitialDelay() const override;
+  double InitialDelay() const override;
   int NextCheckDelay() const override;
   int OnDemandDelay() const override;
   int UpdateDelay() const override;
@@ -108,7 +108,7 @@ ChromeConfigurator::ChromeConfigurator(const base::CommandLine* cmdline,
   DCHECK(pref_service_);
 }
 
-int ChromeConfigurator::InitialDelay() const {
+double ChromeConfigurator::InitialDelay() const {
   return configurator_impl_.InitialDelay();
 }
