+++ chrome/browser/extensions/updater/chrome_update_client_config.cc
@@ -142,7 +142,7 @@ ChromeUpdateClientConfig::ChromeUpdateClientConfig(
   DCHECK(pref_service_);
 }
 
-int ChromeUpdateClientConfig::InitialDelay() const {
+double ChromeUpdateClientConfig::InitialDelay() const {
   return impl_.InitialDelay();
 }
