+++ components/component_updater/configurator_impl.cc
@@ -55,7 +55,7 @@ ConfiguratorImpl::ConfiguratorImpl(
 
 ConfiguratorImpl::~ConfiguratorImpl() = default;
 
-int ConfiguratorImpl::InitialDelay() const {
+double ConfiguratorImpl::InitialDelay() const {
   if (initial_delay_)
     return initial_delay_;
   return fast_update_ ? 10 : kDelayOneMinute;
